+++ src/index.js
@@ -1,7 +1,8 @@
 
-import ReactDOM from 'react-dom'
 import React from 'react'
+import ReactDOM from 'react-dom'
 
+import { loader } from './App';
 import Dev from "./dev";
 
 import * as serviceWorker from './serviceWorker';
@@ -12,66 +13,29 @@ import * as serviceWorker from './serviceWorker';
 
 // import * as queryFragments from "./schema/generated/api.fragments";
 
+
 const node = document.getElementById('root');
 
 if (node) {
 
-  // global.__APOLLO_STATE__
 
-  const render = function () {
-    ReactDOM.render(<Dev
-    // queryFragments={queryFragments}
-    // lang="en"
-    />, node);
-  }
+  loader(node)
+    .then(() => {
 
-  const {
-    __APOLLO_STATE_ID__,
-  } = global;
+      ReactDOM.render(<Dev
+      // queryFragments={queryFragments}
+      // lang="en"
+      />, node);
 
-
-
-  // const uri = new URI();
-
-  // console.log(uri, "uri");
-
-  // const {
-  //   0: base_path,
-  //   1: apollo_state_id,
-  // } = uri.segment();
-
-
-  // if (__APOLLO_STATE_ID__ && base_path === "__apollo-state__" && apollo_state_id) {
-  if (__APOLLO_STATE_ID__) {
-
-    fetch(`/__apollo-state__/${__APOLLO_STATE_ID__}`, {
-      headers: {
-        'Content-Type': 'application/json',
-      },
+      // If you want your app to work offline and load faster, you can change
+      // unregister() to register() below. Note this comes with some pitfalls.
+      // Learn more about service workers: https://bit.ly/CRA-PWA
+      serviceWorker.unregister();
     })
-      .then(r => r.json())
-      .then(state => {
+    .catch(console.error)
 
-        // console.log("state", state);
-
-        global.__APOLLO_STATE__ = state;
-
-        render();
-
-        return;
-      })
-      .catch(console.error)
-
-  }
-  else {
-    render();
-  }
-
-
-
-  // If you want your app to work offline and load faster, you can change
-  // unregister() to register() below. Note this comes with some pitfalls.
-  // Learn more about service workers: https://bit.ly/CRA-PWA
-  serviceWorker.unregister();
 }
+
+
+